ClassifierResult moved to types/
