+++ src/CMakeLists.txt
@@ -83,8 +83,6 @@ set(algorithm_files
     nupic/algorithms/Cell.hpp
     nupic/algorithms/Cells4.cpp
     nupic/algorithms/Cells4.hpp
-    nupic/algorithms/ClassifierResult.cpp
-    nupic/algorithms/ClassifierResult.hpp
     nupic/algorithms/Connections.cpp
     nupic/algorithms/Connections.hpp
     nupic/algorithms/InSynapse.cpp
@@ -195,6 +193,8 @@ set(regions_files
 set(types_files
     nupic/types/BasicType.cpp
     nupic/types/BasicType.hpp
+    nupic/types/ClassifierResult.cpp
+    nupic/types/ClassifierResult.hpp
     nupic/types/Exception.hpp
     nupic/types/Serializable.hpp
     nupic/types/Types.hpp
