+++ src/nupic/types/ClassifierResult.cpp
@@ -21,18 +21,13 @@
  */
 
 #include <cmath>
-#include <map>
-#include <vector>
 
-#include <nupic/algorithms/ClassifierResult.hpp>
-#include <nupic/types/Types.hpp>
+#include <nupic/types/ClassifierResult.hpp>
 #include <nupic/utils/Log.hpp>
 
-using namespace std;
 
 namespace nupic {
-namespace algorithms {
-namespace cla_classifier {
+namespace types {
 
 ClassifierResult::~ClassifierResult() {
   for (map<Int, vector<Real64> *>::const_iterator it = result_.begin();
@@ -43,8 +38,7 @@ ClassifierResult::~ClassifierResult() {
 
 vector<Real64> *ClassifierResult::createVector(Int step, UInt size,
                                                Real64 value) {
-  NTA_CHECK(result_.count(step) == 0)
-      << "The ClassifierResult cannot be reused!";
+  NTA_CHECK(result_.count(step) == 0) << "The ClassifierResult cannot be reused!";
   vector<Real64> *v = new vector<Real64>(size, value);
   result_.insert(pair<Int, vector<Real64> *>(step, v));
   return v;
@@ -66,6 +60,5 @@ bool ClassifierResult::operator==(const ClassifierResult &other) const {
   return true;
 }
 
-} // end namespace cla_classifier
 } // end namespace algorithms
 } // end namespace nupic
